feat: submission trends and tanstack devtool
