+++ frontend/package.json
@@ -11,6 +11,7 @@
     "@carbon/pictograms-react": "^11.69.0",
     "@carbon/react": "^1.71.1",
     "@tanstack/react-query": "^5.50.1",
+    "@tanstack/react-query-devtools": "^5.66.0",
     "@types/node": "^22.0.0",
     "@vitejs/plugin-react": "^4.0.4",
     "@vitejs/plugin-react-swc": "^3.3.2",
