+++ frontend/src/App.tsx
@@ -1,4 +1,7 @@
 import { createBrowserRouter, Navigate, type RouteObject, RouterProvider } from 'react-router-dom';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import { isAxiosError } from 'axios';
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import Landing from "./screens/Landing";
 import SideLayout from './layouts/SideLayout';
 import ProtectedRoute from './routes/ProtectedRoute';
@@ -7,6 +10,7 @@ import SilvicultureSearch from './screens/SilvicultureSearch';
 import ErrorHandling from './screens/ErrorHandling';
 import { useAuth } from './contexts/AuthProvider';
 import { Loading } from '@carbon/react';
+import { THREE_HOURS } from './config/TimeUnits';
 
 import './styles/theme.scss';
 import './styles/default-components.scss'
@@ -48,6 +52,33 @@ const protectedRoutes: RouteObject[] = [
   }
 ];
 
+// Tanstack Query Config
+const HTTP_STATUS_TO_NOT_RETRY = [400, 401, 403, 404];
+const MAX_RETRIES = 3;
+const queryClient = new QueryClient(
+  {
+    defaultOptions: {
+      queries: {
+        refetchOnMount: false, // Default is caching fetched values
+        refetchOnWindowFocus: false,
+        staleTime: THREE_HOURS,
+        gcTime: THREE_HOURS,
+        // Do not retry on errors defined above
+        retry: (failureCount, error) => {
+          if (failureCount > MAX_RETRIES) {
+            return false;
+          }
+          if (
+            isAxiosError(error)
+            && HTTP_STATUS_TO_NOT_RETRY.includes(error.response?.status ?? 0)
+          ) {
+            return false;
+          }
+          return true;
+        }
+      }
+    }
+  });
 
 const App: React.FC = () => {
   const auth = useAuth();
@@ -58,7 +89,12 @@ const App: React.FC = () => {
 
   const browserRouter = createBrowserRouter(auth.isLoggedIn ? protectedRoutes : publicRoutes);
 
-  return <RouterProvider router={browserRouter} />;
+  return (
+    <QueryClientProvider client={queryClient}>
+      <ReactQueryDevtools initialIsOpen={false} />
+      <RouterProvider router={browserRouter} />
+    </QueryClientProvider>
+  );
 };
 
 export default App;
