+++ frontend/src/__test__/components/OpeningMetricsTab.test.tsx
@@ -8,7 +8,7 @@ import { BrowserRouter } from "react-router-dom";
 
 // Utility functions
 import { fetchOpeningFavourites } from '../../services/OpeningFavouriteService';
-import { fetchFreeGrowingMilestones, fetchOpeningsPerYear, fetchRecentOpenings, fetchRecentActions } from '../../services/OpeningService';
+import { fetchFreeGrowingMilestones, fetchUserSubmissionTrends, fetchRecentOpenings, fetchRecentActions } from '../../services/OpeningService';
 
 // Local components
 import OpeningMetricsTab from '../../components/OpeningMetricsTab';
@@ -24,7 +24,7 @@ vi.mock('../../services/OpeningService', async () => {
   return {
     ...actual,
     fetchRecentOpenings: vi.fn(),
-    fetchOpeningsPerYear: vi.fn(),
+    fetchUserSubmissionTrends: vi.fn(),
     fetchFreeGrowingMilestones: vi.fn(),
     fetchRecentActions: vi.fn(),
   };
@@ -34,23 +34,23 @@ describe('OpeningMetricsTab', () => {
   beforeEach(() => {
     vi.clearAllMocks();
     (fetchRecentOpenings as vi.Mock).mockResolvedValue([]);
-    (fetchOpeningsPerYear as vi.Mock).mockResolvedValue([]);
+    (fetchUserSubmissionTrends as vi.Mock).mockResolvedValue([]);
     (fetchFreeGrowingMilestones as vi.Mock).mockResolvedValue([{ group: '1-5', value: 11 }]);
     (fetchOpeningFavourites as vi.Mock).mockResolvedValue([1, 2, 3]);
     (fetchRecentActions as vi.Mock).mockResolvedValue([
       {
-        activityType: "Update",        
+        activityType: "Update",
         openingId: "1541297",
         statusCode: "APP",
         lastUpdated: "2024-05-16T19:59:21.635Z",
         lastUpdatedLabel: "1 minute ago"
       }
     ]);
-    
+
   });
 
   it('should render the OpeningMetricsTab component with all sections', async () => {
-    
+
     await act(async () => render(<BrowserRouter><NotificationProvider><OpeningMetricsTab /></NotificationProvider></BrowserRouter>));
 
     expect(screen.getByText('Dashboard')).toBeInTheDocument();
@@ -64,42 +64,42 @@ describe('OpeningMetricsTab', () => {
   });
 
   it('should call fetchOpeningFavourites and set submissionTrends state', async () => {
-    
+
     let container;
     await act(async () => {
       ({ container } = render(<BrowserRouter><NotificationProvider><OpeningMetricsTab /></NotificationProvider></BrowserRouter>));
     });
 
     await waitFor(() => {
-      expect(fetchOpeningFavourites).toHaveBeenCalled();      
+      expect(fetchOpeningFavourites).toHaveBeenCalled();
     });
 
     expect(container).not.toBeNull();
-    
+
     expect(container.querySelector('.row.activity-history-container.gx-4')).toBeInTheDocument();
-    
+
     expect(
       container
-      .querySelector('.row.activity-history-container.gx-4')
-      .innerHTML
+        .querySelector('.row.activity-history-container.gx-4')
+        .innerHTML
     ).toContain('<span class="trend-title">Opening ID</span>&nbsp;1');
-    
+
     expect(
       container
-      .querySelector('.row.activity-history-container.gx-4')
-      .innerHTML
+        .querySelector('.row.activity-history-container.gx-4')
+        .innerHTML
     ).toContain('<span class="trend-title">Opening ID</span>&nbsp;2');
-    
+
     expect(
       container
-      .querySelector('.row.activity-history-container.gx-4')
-      .innerHTML
+        .querySelector('.row.activity-history-container.gx-4')
+        .innerHTML
     ).toContain('<span class="trend-title">Opening ID</span>&nbsp;3');
 
   });
 
   it('should scroll to "Track Openings" section when scrollTo parameter is "trackOpenings"', async () => {
-    
+
     const mockScrollIntoView = vi.fn();
     window.HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
 
@@ -115,7 +115,7 @@ describe('OpeningMetricsTab', () => {
   });
 
   it('should not scroll to "Track Openings" section when scrollTo parameter is not "trackOpenings"', async () => {
-    
+
     const mockScrollIntoView = vi.fn();
     window.HTMLElement.prototype.scrollIntoView = mockScrollIntoView;
