+++ frontend/src/__test__/screens/Opening.test.tsx
@@ -6,7 +6,7 @@ import PaginationContext from '../../contexts/PaginationContext';
 import { NotificationProvider } from '../../contexts/NotificationProvider';
 import { BrowserRouter } from 'react-router-dom';
 import { RecentOpening } from '../../types/RecentOpening';
-import { fetchFreeGrowingMilestones, fetchOpeningsPerYear, fetchRecentActions } from '../../services/OpeningService';
+import { fetchFreeGrowingMilestones, fetchUserSubmissionTrends, fetchRecentActions } from '../../services/OpeningService';
 import { fetchOpeningFavourites } from '../../services/OpeningFavouriteService';
 import { AuthProvider } from '../../contexts/AuthProvider';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
@@ -28,7 +28,7 @@ vi.mock('../../services/OpeningService', async () => {
   const actual = await vi.importActual('../../services/OpeningService');
   return {
     ...actual,
-    fetchOpeningsPerYear: vi.fn(),
+    fetchUserSubmissionTrends: vi.fn(),
     fetchFreeGrowingMilestones: vi.fn(),
     fetchRecentActions: vi.fn(),
   };
@@ -55,7 +55,7 @@ const rows: RecentOpening[] = [{
   entryTimestamp: '1',
   updateTimestamp: '1',
 }];
-  
+
 const paginationValueMock = {
   getCurrentData: () => rows,
   currentPage: 0,
@@ -72,13 +72,13 @@ const queryClient = new QueryClient();
 describe('Opening screen test cases', () => {
 
   beforeEach(() => {
-    vi.clearAllMocks(); 
-    (fetchOpeningsPerYear as vi.Mock).mockResolvedValue([
+    vi.clearAllMocks();
+    (fetchUserSubmissionTrends as vi.Mock).mockResolvedValue([
       { group: '2022', key: 'Openings', value: 10 },
       { group: '2023', key: 'Openings', value: 15 },
     ]);
     (fetchFreeGrowingMilestones as vi.Mock).mockResolvedValue([{ group: '1-5', value: 11 }]);
-    (fetchOpeningFavourites as vi.Mock).mockResolvedValue([1,2,3]);
+    (fetchOpeningFavourites as vi.Mock).mockResolvedValue([1, 2, 3]);
     (fetchRecentActions as vi.Mock).mockResolvedValue([data]);
 
   });
@@ -86,14 +86,14 @@ describe('Opening screen test cases', () => {
   it('should renders Opening Page Title component', async () => {
     const { getByTestId } = render(
       <BrowserRouter>
-      <QueryClientProvider client={queryClient}>
-        <PaginationContext.Provider value={paginationValueMock}>
-          <NotificationProvider>
-            <AuthProvider>
-              <Opening />
-            </AuthProvider>
-          </NotificationProvider>
-        </PaginationContext.Provider>
+        <QueryClientProvider client={queryClient}>
+          <PaginationContext.Provider value={paginationValueMock}>
+            <NotificationProvider>
+              <AuthProvider>
+                <Opening />
+              </AuthProvider>
+            </NotificationProvider>
+          </PaginationContext.Provider>
         </QueryClientProvider>
       </BrowserRouter>
     );
@@ -106,18 +106,18 @@ describe('Opening screen test cases', () => {
   });
 
   it('should render Opening Page with dashboard tab selected', async () => {
-    window.history.pushState({}, 'Opening page with metrics','/opening?tab=metrics');
+    window.history.pushState({}, 'Opening page with metrics', '/opening?tab=metrics');
 
     await act(async () => render(
       <BrowserRouter>
-      <QueryClientProvider client={queryClient}>
-        <PaginationContext.Provider value={paginationValueMock}>
-          <NotificationProvider>
-            <AuthProvider>
-              <Opening />
-            </AuthProvider>
-          </NotificationProvider>
-        </PaginationContext.Provider>
+        <QueryClientProvider client={queryClient}>
+          <PaginationContext.Provider value={paginationValueMock}>
+            <NotificationProvider>
+              <AuthProvider>
+                <Opening />
+              </AuthProvider>
+            </NotificationProvider>
+          </PaginationContext.Provider>
         </QueryClientProvider>
       </BrowserRouter>
     ));
@@ -129,34 +129,34 @@ describe('Opening screen test cases', () => {
   describe('FavoriteCards test cases', () => {
 
     it('should render FavoriteCard component', async () => {
-      
+
       let container: HTMLElement = document.createElement('div');
       await act(async () => {
-      ({ container } = render(
-        <BrowserRouter>
-          <QueryClientProvider client={queryClient}>
-            <PaginationContext.Provider value={paginationValueMock}>
-              <NotificationProvider>
-                <AuthProvider>
-                  <Opening />
-                </AuthProvider>
-              </NotificationProvider>
-            </PaginationContext.Provider>
-          </QueryClientProvider>
-        </BrowserRouter>
-      ));
-    });
-      
-      // check if first FavoriteCard has the correct title and is active      
+        ({ container } = render(
+          <BrowserRouter>
+            <QueryClientProvider client={queryClient}>
+              <PaginationContext.Provider value={paginationValueMock}>
+                <NotificationProvider>
+                  <AuthProvider>
+                    <Opening />
+                  </AuthProvider>
+                </NotificationProvider>
+              </PaginationContext.Provider>
+            </QueryClientProvider>
+          </BrowserRouter>
+        ));
+      });
+
+      // check if first FavoriteCard has the correct title and is active
       expect(container.querySelector('#fav-card-1')).toBeDefined();
       expect(container.querySelector('#fav-card-1')?.textContent).toContain('Silviculture search');
       expect(container.querySelector('#fav-card-1')?.className).not.contain('cds--link--disable');
-      
+
       // check if the second FavoriteCard has the correct title and is inactive
       expect(container.querySelector('#fav-card-2')).toBeDefined();
       expect(container.querySelector('#fav-card-2')?.textContent).toContain('Create an opening');
       expect(container.querySelector('#fav-card-2')?.className).toContain('cds--link--disable');
-      
+
       // check if the third FavoriteCard has the correct title and is inactive
       expect(container.querySelector('#fav-card-3')).toBeDefined();
       expect(container.querySelector('#fav-card-3')?.textContent).toContain('Reports');
@@ -170,24 +170,24 @@ describe('Opening screen test cases', () => {
     });
 
     it('should not render tab when not selected', async () => {
-      window.history.pushState({}, 'Opening page with metrics','/opening?tab=openings');
+      window.history.pushState({}, 'Opening page with metrics', '/opening?tab=openings');
       let container: HTMLElement = document.createElement('div');
       let getByText: any;
       await act(async () => {
-      ({ container, getByText } = render(
-        <BrowserRouter>
-          <QueryClientProvider client={queryClient}>
-            <PaginationContext.Provider value={paginationValueMock}>
-              <NotificationProvider>
-                <AuthProvider>
-                  <Opening />
-                </AuthProvider>
-              </NotificationProvider>
-            </PaginationContext.Provider>
-          </QueryClientProvider>
-        </BrowserRouter>
-      ));
-    });
+        ({ container, getByText } = render(
+          <BrowserRouter>
+            <QueryClientProvider client={queryClient}>
+              <PaginationContext.Provider value={paginationValueMock}>
+                <NotificationProvider>
+                  <AuthProvider>
+                    <Opening />
+                  </AuthProvider>
+                </NotificationProvider>
+              </PaginationContext.Provider>
+            </QueryClientProvider>
+          </BrowserRouter>
+        ));
+      });
 
       // check if the tab is not rendered
       expect(container.querySelector('div.tab-openings')?.childNodes).toHaveLength(2);
@@ -196,25 +196,25 @@ describe('Opening screen test cases', () => {
     });
 
     it('should render tab only when selected', async () => {
-      window.history.pushState({}, 'Opening page with metrics','/opening');
+      window.history.pushState({}, 'Opening page with metrics', '/opening');
 
       let container: HTMLElement = document.createElement('div');
       let getByText: any;
       await act(async () => {
-      ({ container, getByText } = render(
-        <BrowserRouter>
-          <QueryClientProvider client={queryClient}>
-            <PaginationContext.Provider value={paginationValueMock}>
-              <NotificationProvider>
-                <AuthProvider>
-                  <Opening />
-                </AuthProvider>
-              </NotificationProvider>
-            </PaginationContext.Provider>
-          </QueryClientProvider>
-        </BrowserRouter>
-      ));
-    });
+        ({ container, getByText } = render(
+          <BrowserRouter>
+            <QueryClientProvider client={queryClient}>
+              <PaginationContext.Provider value={paginationValueMock}>
+                <NotificationProvider>
+                  <AuthProvider>
+                    <Opening />
+                  </AuthProvider>
+                </NotificationProvider>
+              </PaginationContext.Provider>
+            </QueryClientProvider>
+          </BrowserRouter>
+        ));
+      });
 
       await act(async () => getByText('Dashboard').click());
