--- frontend/src/components/BarChartGrouped/BarChartGrouped.scss
@@ -1,16 +1,0 @@
-@use '@bcgov-nr/nr-theme/design-tokens/colors.scss' as colors;
-@use '@bcgov-nr/nr-theme/design-tokens/variables.scss' as vars;
-
-.cds--cc--axes g.axis g.tick text {
-    fill: var(--#{vars.$bcgov-prefix}-text-secondary);
-}
-.cds--cc--grid rect.chart-grid-backdrop {
-    fill: var(--#{vars.$bcgov-prefix}-layer-02);
-}
-.#{vars.$bcgov-prefix}--date-picker-container {
- min-height: 66px;
-}
-.bar-chart-container {
-    padding-top: 1rem;
-    padding-bottom: 1rem;
-}
