+++ frontend/src/components/ChartContainer/index.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
-import './ChartContainer.scss'
+import { Column, Grid } from '@carbon/react';
 import ChartTitle from '../ChartTitle';
+
+import './styles.scss'
 
 type Props = {
   children?: React.ReactNode;
@@ -19,17 +21,12 @@ type Props = {
  */
 function ChartContainer({ children, title, description }: Props): JSX.Element {
   return (
-    <div className='chart-container'>
-        <div className="d-flex flex-row">
-          <ChartTitle title={title} subtitle={description} />
-        </div>
-        {children? (
-          <div className="content">
-            {children}
-          </div>
-        ) : null}
-        
-    </div>
+    <Grid className='chart-container-grid'>
+      <Column sm={4} md={8} lg={16}>
+        <ChartTitle title={title} subtitle={description} />
+      </Column>
+      {children}
+    </Grid>
   );
 }
