+++ frontend/src/components/ChartTitle/index.tsx
@@ -1,5 +1,6 @@
-import { Column} from '@carbon/react';
+import { Column } from '@carbon/react';
 import Subtitle from '../Subtitle';
+
 import './styles.scss';
 
 interface ChartTitleProps {
@@ -19,7 +20,7 @@ interface ChartTitleProps {
  * @param {string} [props.activity] - The activity of the chart.
  * @returns {JSX.Element} A div element containing the chart title.
  */
-function ChartTitle ({ title, subtitle}: ChartTitleProps) {
+function ChartTitle({ title, subtitle }: ChartTitleProps) {
   return (
     <Column className="chart-title">
       <div className="title-favourite">
