+++ frontend/src/components/ChartTitle/styles.scss
@@ -2,7 +2,6 @@
 
 .chart-title {
   margin: 0;
-  padding: 16px;
 }
 
 .chart-title h1 {
@@ -14,4 +13,3 @@
   display: flex;
   align-items: center;
 }
-
