+++ frontend/src/components/FavouriteCard/styles.scss
@@ -6,12 +6,14 @@
   height: 4rem;
   border: 1px solid var(--#{vars.$bcgov-prefix}-border-subtle-02);
   background-color: var(--#{vars.$bcgov-prefix}-layer-02);
+  min-width: fit-content;
 
   .fav-card-content {
     display: flex;
     flex-direction: row;
     align-items: center;
     gap: 0.75rem;
+    white-space: nowrap;
   }
 
   .fav-card-icon {
