+++ frontend/src/components/OpeningSubmissionTrend/definitions.ts
@@ -1,0 +1,10 @@
+import { OpeningsPerYearDto } from "../../types/OpeningTypes";
+
+/**
+ * The type that is used by the chart.
+ */
+export type SubmissionTrendChartObj = OpeningsPerYearDto & {
+  group: string;
+  key: string;
+  value: number;
+}
