+++ frontend/src/components/RecentOpenings/styles.scss
@@ -34,7 +34,6 @@
     display: flex;
     flex-direction: row;
     align-items: center;
-    justify-content: center;
   }
 
   @media only screen and (max-width: 671px) {
