+++ frontend/src/index.tsx
@@ -19,33 +19,10 @@ import { NotificationProvider } from './contexts/NotificationProvider';
 const container: HTMLElement | null = document.getElementById('root');
 if (container) {
   const root = createRoot(container);
-  const HTTP_STATUS_TO_NOT_RETRY = [400, 401, 403, 404];
-  const MAX_RETRIES = 3;
-  const queryClient = new QueryClient(
-    {
-      defaultOptions: {
-        queries: {
-          refetchOnMount: false,
-          refetchOnWindowFocus: false,
-          // Do not retry on errors defined above
-          retry: (failureCount, error) => {
-            if (failureCount > MAX_RETRIES) {
-              return false;
-            }
-            if (
-              isAxiosError(error)
-              && HTTP_STATUS_TO_NOT_RETRY.includes(error.response?.status ?? 0)
-            ) {
-              return false;
-            }
-            return true;
-          }
-        }
-      }
-    });
 
-    Amplify.configure(amplifyconfig);
-    cognitoUserPoolsTokenProvider.setKeyValueStorage(new CookieStorage());
+
+  Amplify.configure(amplifyconfig);
+  cognitoUserPoolsTokenProvider.setKeyValueStorage(new CookieStorage());
 
 
   root.render(
@@ -53,15 +30,13 @@ if (container) {
       <ClassPrefix prefix='bx'>
         <ThemePreference>
           <AuthProvider>
-            <QueryClientProvider client={queryClient}>        
-                  <PaginationProvider>
-                    <OpeningsSearchProvider>
-                      <NotificationProvider>
-                        <App />
-                      </NotificationProvider>
-                    </OpeningsSearchProvider>
-                  </PaginationProvider>
-            </QueryClientProvider>
+            <PaginationProvider>
+              <OpeningsSearchProvider>
+                <NotificationProvider>
+                  <App />
+                </NotificationProvider>
+              </OpeningsSearchProvider>
+            </PaginationProvider>
           </AuthProvider>
         </ThemePreference>
       </ClassPrefix>
