--- frontend/src/types/OpeningPerYearChart.ts
@@ -1,8 +1,0 @@
-export interface OpeningPerYearChart {
-  group: string;
-  key: string;
-  value: string;
-  year: number;
-  month: number;
-  statusCount: Record<string, number>;
-}
